Fix INTEGER overflow on numero_contrato cast (Cierre Diario detail)
Sixt rental numbers like 9523073821 (~9.5B) exceed Postgres INTEGER max
(2.1B). SQLite was forgiving since INTEGER there is dynamic 64-bit, but
Postgres INTEGER is strict 32-bit. Cast to BIGINT instead.

+++ dashboard_v2/pages/1_Cierre_Diario.py
@@ -181,7 +181,7 @@ else:
     FROM vw_rentals_detail
     WHERE fuente_cargo = 'RENTAL_COUNTER'
       AND rental_currency = 'USD'
-      AND CAST(numero_contrato AS INTEGER) IN ({placeholders})
+      AND CAST(numero_contrato AS BIGINT) IN ({placeholders})
     ORDER BY DATE(fecha_handover_real) DESC, numero_contrato,
              cargo_inty, cargo_posicion
     """
